+++ R/zzz.R
@@ -7,7 +7,8 @@
     .env = ns,
     vec_interval_groups = import_vctrs("exp_vec_interval_groups"),
     vec_interval_locate_groups = import_vctrs("exp_vec_interval_locate_groups"),
-    vec_interval_complement = import_vctrs("exp_vec_interval_complement")
+    vec_interval_complement = import_vctrs("exp_vec_interval_complement"),
+    vec_interval_locate_containers = import_vctrs("exp_vec_interval_locate_containers")
   )
 }