+++ NEWS.md
@@ -1,5 +1,10 @@
 # ivs 0.1.0.9000
 
+* New family of functions for identifying interval _containers_, which are
+  intervals that aren't contained within any other interval. These functions
+  are: `iv_containers()`, `iv_identify_containers()`, `iv_identify_container()`,
+  and `iv_locate_containers()` (#20).
+
 * New Examples vignette that links out to Stack Overflow questions that were
   solved with ivs. View it locally with `vignette("examples", package = "ivs")`.